requirements.txt
datasets
sentence-transformers
faiss-cpu
openai
python-dotenv
torch